Move legacy functionality into SimpleFormsLegacy
