+++ composer.json
@@ -1,42 +1,44 @@
 {
-    "name": "bolt/simpleforms",
-    "description": "This extension will allow you to insert simple forms on your site, for users to get in touch, send you a quick note or something like that. To use, configure the required fields in config.yml, and place <code>{{ simpleform('contact') }}</code> in your templates.",
-    "type": "bolt-extension",
-    "keywords": [
-        "forms",
-        "input",
-        "email",
-        "upload"
-    ],
-    "require": {
-        "bolt/bolt": ">=2.0.0,<3.0.0"
-    },
-    "authors": [
-        {
-            "name": "Lodewijk Evers",
-            "email": "lodewijk@twokings.nl",
-            "homepage": "https://github.com/jadwigo"
-        }
-    ],
-    "minimum-stability" : "dev",
-    "prefer-stable" : true,
-    "autoload": {
-        "files": [
-            "init.php"
-        ],
-        "psr-4" : {
-            "Bolt\\Extension\\Bolt\\SimpleForms\\" : ""
-        }
-    },
-    "autoload-dev" : {
-         "psr-4" : {
-            "Bolt\\Extension\\Bolt\\SimpleForms\\Tests\\" : "tests/",
-            "Bolt\\Tests\\" : "vendor/bolt/bolt/tests/phpunit/unit/"
-         }
-     },
-    "extra": {
-        "branch-alias": {
-            "dev-master" : "2.0.*-dev"
-        }
-    }
+	"name" : "bolt/simpleforms",
+	"description" : "This extension will allow you to insert simple forms on your site, for users to get in touch, send you a quick note or something like that. To use, configure the required fields in config.yml, and place <code>{{ simpleform('contact') }}</code> in your templates.",
+	"type" : "bolt-extension",
+	"keywords" : [
+		"forms",
+		"input",
+		"email",
+		"upload"
+	],
+	"require" : {
+		"bolt/bolt" : ">=2.0.0,<3.0.0"
+	},
+	"authors" : [{
+			"name" : "Lodewijk Evers",
+			"email" : "lodewijk@twokings.nl",
+			"homepage" : "https://github.com/jadwigo"
+		}
+	],
+	"minimum-stability" : "dev",
+	"prefer-stable" : true,
+	"autoload" : {
+		"files" : [
+			"init.php"
+		],
+		"psr-4" : {
+			"Bolt\\Extension\\Bolt\\SimpleForms\\" : [
+				"",
+				"src"
+			]
+		}
+	},
+	"autoload-dev" : {
+		"psr-4" : {
+			"Bolt\\Extension\\Bolt\\SimpleForms\\Tests\\" : "tests/",
+			"Bolt\\Tests\\" : "vendor/bolt/bolt/tests/phpunit/unit/"
+		}
+	},
+	"extra" : {
+		"branch-alias" : {
+			"dev-master" : "2.0.*-dev"
+		}
+	}
 }
